fix: reduced bugs and stable version

+++ app/main/io.js
@@ -5,12 +5,11 @@ const os = require("os");
 const open = require("open");
 const chokidar = require("chokidar");
 
-
 // local dependencies
 const notification = require("./notification");
 
 // get application directory
-const appDir = path.resolve(__dirname+"/python", "src");
+const appDir = path.resolve(__dirname + "/python", "src");
 
 /****************************/
 
@@ -30,6 +29,13 @@ exports.getFiles = () => {
   });
 };
 
+/// return the number of files added to the wizard
+exports.getFileCount = () => {
+  const files = fs.readdirSync(appDir);
+
+  return files.length;
+};
+
 /****************************/
 
 // add files
@@ -40,7 +46,6 @@ exports.addFiles = (files = []) => {
   // copy `files` recursively (ignore duplicate file names)
   files.forEach((file) => {
     const filePath = path.resolve(appDir, file.name);
-
     if (!fs.existsSync(filePath)) {
       fs.copyFileSync(file.path, filePath);
     }
@@ -55,9 +60,20 @@ exports.deleteFile = (filename) => {
   const filePath = path.resolve(appDir, filename);
 
   // remove file from the file system
-  if (fs.existsSync(filePath)) {
-    fs.removeSync(filePath);
-  }
+  if (fs.existsSync(filePath)) fs.removeSync(filePath);
+};
+// clearing the output file
+exports.emptyFile = (filePath) => {
+  if (fs.existsSync(filePath))
+    fs.writeFile(filePath, "", () => console.log("output removed"));
+};
+
+/// adding a loader in output file
+exports.addLoader = (filePath) => {
+  if (fs.existsSync(filePath))
+    fs.writeFile(filePath, '<div class="loader"></div>', () =>
+      console.log("done")
+    );
 };
 
 // open a file
@@ -65,9 +81,7 @@ exports.openFile = (filename) => {
   const filePath = path.resolve(appDir, filename);
 
   // open a file using default application
-  if (fs.existsSync(filePath)) {
-    open(filePath);
-  }
+  if (fs.existsSync(filePath)) open(filePath);
 };
 
 /*-----*/
